DCS-12: Some file moving and replacement.
